requirements.txt
asgiref==3.8.1
Django==5.0.7
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
